requirements.txt
kafka-python==2.0.2
python-socketio==5.11.0
asyncio-mqtt==0.16.1
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0